requirements.txt
accelerate==1.5.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.12.3
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.2
confection==0.1.4
cymem==2.0.8
datasets==3.3.2
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
distro==1.9.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.28.1
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
langcodes==3.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.65.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
propcache==0.3.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.2
transformers==4.49.0
triton==3.2.0
typer==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
xxhash==3.5.0
yarl==1.18.3
